6. move inner class to upper
6. move inner class to upper
