+++ moim/src/test/java/com/youth/moim/presentation/user/Gender.java
@@ -1,0 +1,5 @@
+package com.youth.moim.presentation.user;
+
+public enum Gender {
+    FEMALE, MALE
+}
